+++ Domain/Domain.csproj
@@ -9,5 +9,8 @@
   <ItemGroup>
     <PackageReference Include="MediatR" Version="12.5.0" />
   </ItemGroup>
+	<ItemGroup>
+		<InternalsVisibleTo Include="Application.Test" />
+	</ItemGroup>
 	
 </Project>
