+++ TaskManagerApi.sln
@@ -19,7 +19,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastru
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebAPI", "WebAPI\WebAPI.csproj", "{8761813D-16B2-4A88-87E3-5AC2E4EA5A34}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application.Test", "Application.Test\Application.Test.csproj", "{5071B474-1625-4005-84EF-30407E98ABA7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application.Test", "Application.Tests\Application.Test.csproj", "{F42A95DA-6F13-6D71-6074-2F812D48CCD5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,10 +43,10 @@ Global
 		{8761813D-16B2-4A88-87E3-5AC2E4EA5A34}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8761813D-16B2-4A88-87E3-5AC2E4EA5A34}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8761813D-16B2-4A88-87E3-5AC2E4EA5A34}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5071B474-1625-4005-84EF-30407E98ABA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5071B474-1625-4005-84EF-30407E98ABA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5071B474-1625-4005-84EF-30407E98ABA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5071B474-1625-4005-84EF-30407E98ABA7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F42A95DA-6F13-6D71-6074-2F812D48CCD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F42A95DA-6F13-6D71-6074-2F812D48CCD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F42A95DA-6F13-6D71-6074-2F812D48CCD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F42A95DA-6F13-6D71-6074-2F812D48CCD5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
